lerobot/franka_sim/requirements.txt
dm_env
mujoco==3.2.6
gym >= 0.26
gymnasium
dm-robotics-transformations
imageio[ffmpeg]
lxml
inputs==0.5
